Remove PinotColumn
Remove PinotColumn and moved used methods to PinotColumnHandle instead
