+++ plugin/trino-pinot/src/main/java/io/trino/plugin/pinot/PinotColumnHandle.java
@@ -17,11 +17,26 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.trino.spi.connector.ColumnHandle;
 import io.trino.spi.connector.ColumnMetadata;
+import io.trino.spi.type.ArrayType;
+import io.trino.spi.type.BigintType;
+import io.trino.spi.type.BooleanType;
+import io.trino.spi.type.DoubleType;
+import io.trino.spi.type.IntegerType;
+import io.trino.spi.type.RealType;
 import io.trino.spi.type.Type;
+import io.trino.spi.type.VarbinaryType;
+import io.trino.spi.type.VarcharType;
+import org.apache.pinot.core.operator.transform.TransformResultMetadata;
+import org.apache.pinot.spi.data.FieldSpec;
+import org.apache.pinot.spi.data.Schema;
 
+import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 import static com.google.common.base.MoreObjects.toStringHelper;
+import static com.google.common.collect.ImmutableList.toImmutableList;
+import static io.trino.plugin.pinot.PinotErrorCode.PINOT_UNSUPPORTED_COLUMN_TYPE;
 import static java.util.Objects.requireNonNull;
 
 public class PinotColumnHandle
@@ -45,6 +60,57 @@ public class PinotColumnHandle
         this.columnName = requireNonNull(columnName, "columnName is null");
         this.dataType = requireNonNull(dataType, "dataType is null");
         this.returnNullOnEmptyGroup = returnNullOnEmptyGroup;
+    }
+
+    public static List<PinotColumnHandle> getPinotColumnsForPinotSchema(Schema pinotTableSchema)
+    {
+        return pinotTableSchema.getColumnNames().stream()
+                .filter(columnName -> !columnName.startsWith("$")) // Hidden columns starts with "$", ignore them as we can't use them in PQL
+                .map(columnName -> new PinotColumnHandle(columnName, getTrinoTypeFromPinotType(pinotTableSchema.getFieldSpecFor(columnName))))
+                .collect(toImmutableList());
+    }
+
+    public static Type getTrinoTypeFromPinotType(FieldSpec field)
+    {
+        Type type = getTrinoTypeFromPinotType(field.getDataType());
+        if (field.isSingleValueField()) {
+            return type;
+        }
+        else {
+            return new ArrayType(type);
+        }
+    }
+
+    public static Type getTrinoTypeFromPinotType(TransformResultMetadata transformResultMetadata)
+    {
+        Type type = getTrinoTypeFromPinotType(transformResultMetadata.getDataType());
+        if (transformResultMetadata.isSingleValue()) {
+            return type;
+        }
+        return new ArrayType(type);
+    }
+
+    public static Type getTrinoTypeFromPinotType(FieldSpec.DataType dataType)
+    {
+        switch (dataType) {
+            case BOOLEAN:
+                return BooleanType.BOOLEAN;
+            case FLOAT:
+                return RealType.REAL;
+            case DOUBLE:
+                return DoubleType.DOUBLE;
+            case INT:
+                return IntegerType.INTEGER;
+            case LONG:
+                return BigintType.BIGINT;
+            case STRING:
+                return VarcharType.VARCHAR;
+            case BYTES:
+                return VarbinaryType.VARBINARY;
+            default:
+                break;
+        }
+        throw new PinotException(PINOT_UNSUPPORTED_COLUMN_TYPE, Optional.empty(), "Unsupported type conversion for pinot data type: " + dataType);
     }
 
     @JsonProperty
